core: move document repo into shared
