+++ core/src/model/mod.rs
@@ -1,3 +1,2 @@
 pub mod drawing;
 pub mod errors;
-pub mod repo;
