+++ core/src/repo/mod.rs
@@ -1,3 +1,1 @@
-pub mod document_repo;
-pub mod local_storage;
 pub mod schema;
